requirements.txt
Django~=4.2
djangorestframework~=3.15.2
django-cors-headers~=
psycopg2-binary~=2.9.9
pillow~=10.0.1
requests~=2.32.0
python-dotenv~=1.0.1
